+++ experimental/packages/opentelemetry-instrumentation-fetch/test/fetch.test.ts
@@ -900,4 +900,20 @@ describe('fetch', () => {
       );
     });
   });
+
+  describe('when network events are ignored', () => {
+    beforeEach(async () => {
+      await prepareData(url, {
+        ignoreNetworkEvents: true,
+      });
+    });
+    afterEach(() => {
+      clearData();
+    });
+    it('should NOT add network events', () => {
+      const span: tracing.ReadableSpan = exportSpy.args[1][0][0];
+      const events = span.events;
+      assert.strictEqual(events.length, 0, 'number of events is wrong');
+    });
+  });
 });