+++ experimental/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :rocket: (Enhancement)
 
+* feat(opentelemetry-instrumentation-fetch): optionally ignore network events #3028 @gregolsen
+
 ### :bug: (Bug Fix)
 
 * fix(otlp-transformer): remove type dependency on Long #3022 @legendecas